feat(DTFS2-8016): gef fixed fee calculation as per guarantee fee
